+++ src/i9950-printer-app.c
@@ -8,6 +8,11 @@
 #include "pappl/i9950_driver.h"
 
 #include <pappl/pappl.h>
+#include <stdlib.h>
+
+#ifdef __APPLE__
+# include "macos/i9950_status_bar.h"
+#endif
 
 #ifndef I9950_VERSION
 #define I9950_VERSION "0.0.0-dev"
@@ -29,6 +34,14 @@ i9950_callback(pappl_system_t         *system,
 int
 main(int argc, char *argv[])
 {
+#ifdef __APPLE__
+  if (i9950_status_bar_wanted(argc, argv))
+  {
+    setenv("I9950_CUSTOM_MENU_BAR", "1", 1);
+    i9950_status_bar_init(i9950_menu_icon_path(), 0);
+  }
+#endif
+
   return papplMainloop(argc, argv,
                        I9950_VERSION,
                        "Copyright (C) 2026 i9950 driver project. "